ChatGPT_Prompt_Engineering_for_Developers/requirements.txt
groq
jupyter
notebook
python-dotenv